refactor: plan-map 훅 분리 작업 (#164)
